// src/content-type.ts
export enum ContentType {
    CREATE_PAYMENT_BANK_V1 = "application/vnd.payments.create-payment.bank.v1+json",
    CREATE_PAYMENT_BANK_RESPONSE_V1 = "application/vnd.payments.create-payment.bank.response.v1+json",
    CREATE_PAYMENT_CARD_V1 = "application/vnd.payments.create-payment.card.v1+json",
    CREATE_PAYMENT_CARD_RESPONSE_V1 = "application/vnd.payments.create-payment.card.response.v1+json",
    CREATE_PAYMENT_PAYMENT_METHOD_V1 = "application/vnd.payments.create-payment.payment-method.v1+json",
    CREATE_PAYMENT_PAYMENT_METHOD_RESPONSE_V1 = "application/vnd.payments.create-payment.payment-method.response.v1+json",
    CREATE_PAYMENT_METHOD_BANK_V1 = "application/vnd.payments.create-payment-method.bank.v1+json",
    CREATE_PAYMENT_METHOD_BANK_RESPONSE_V1 = "application/vnd.payments.create-payment-method.bank.response.v1+json",
    CREATE_PAYMENT_METHOD_CARD_V1 = "application/vnd.payments.create-payment-method.card.v1+json",
    CREATE_PAYMENT_METHOD_CARD_RESPONSE_V1 = "application/vnd.payments.create-payment-method.card.response.v1+json",
    CREATE_PROVIDER_GATEWAY_AUTHORIZE_NET_V1 = "application/vnd.payments.create-provider.gateway.authorize-net.v1+json",
    CREATE_PROVIDER_GATEWAY_AUTHORIZE_NET_RESPONSE_V1 = "application/vnd.ncryptyr.create-provider.gateway.authorize-net.response.v1+json",
    UPDATE_PROVIDER_GATEWAY_AUTHORIZE_NET_V1 = "application/vnd.payments.update-provider.gateway.authorize-net.v1+json",
    UPDATE_PROVIDER_GATEWAY_AUTHORIZE_NET_RESPONSE_V1 = "application/vnd.ncryptyr.update-provider.gateway.authorize-net.response.v1+json",
}